+++ lce/src/main/java/com/laimiux/lce/Type.kt
@@ -7,10 +7,10 @@ sealed class Type<out L, out C, out E> : LCE<L, C, E> {
     /**
      * Loading type has two subclasses [Loading.UnitType] and [Loading.Typed].
      */
-    sealed class Loading<out T>() :
-        Type<T, Nothing, Nothing>(),
-        LCE<T, Nothing, Nothing>,
-        LC<T, Nothing> {
+    sealed class Loading<out L>() :
+        Type<L, Nothing, Nothing>(),
+        LCE<L, Nothing, Nothing>,
+        LC<L, Nothing> {
 
         companion object {
             operator fun invoke() = UnitType
@@ -28,7 +28,7 @@ sealed class Type<out L, out C, out E> : LCE<L, C, E> {
             }
         }
 
-        abstract val value: T
+        abstract val value: L
 
         override fun isLoading(): Boolean = true
         override fun isContent(): Boolean = false
@@ -36,9 +36,9 @@ sealed class Type<out L, out C, out E> : LCE<L, C, E> {
 
         override fun contentOrNull(): Nothing? = null
         override fun errorOrNull(): Nothing? = null
-        override fun loadingOrNull(): T = value
+        override fun loadingOrNull(): L = value
 
-        override fun asLceType(): Type<T, Nothing, Nothing> = this
+        override fun asLceType(): Type<L, Nothing, Nothing> = this
 
         object UnitType : Loading<Unit>(),
             LCE<Unit, Nothing, Nothing>,
@@ -57,11 +57,11 @@ sealed class Type<out L, out C, out E> : LCE<L, C, E> {
     /**
      * Error type has two subclasses [Error.ThrowableType] and [Error.Typed].
      */
-    sealed class Error<out T>() :
-        Type<Nothing, Nothing, T>(),
-        LCE<Nothing, Nothing, T>,
-        UCE<Nothing, T>,
-        CE<Nothing, T> {
+    sealed class Error<out E>() :
+        Type<Nothing, Nothing, E>(),
+        LCE<Nothing, Nothing, E>,
+        UCE<Nothing, E>,
+        CE<Nothing, E> {
 
         companion object {
             operator fun invoke(throwable: Throwable) = ThrowableType(throwable)
@@ -76,17 +76,17 @@ sealed class Type<out L, out C, out E> : LCE<L, C, E> {
             }
         }
 
-        abstract val value: T
+        abstract val value: E
 
         override fun isLoading(): Boolean = false
         override fun isContent(): Boolean = false
         override fun isError(): Boolean = true
 
         override fun contentOrNull(): Nothing? = null
-        override fun errorOrNull(): T = value
+        override fun errorOrNull(): E = value
         override fun loadingOrNull(): Nothing? = null
 
-        override fun asLceType(): Type<Nothing, Nothing, T> = this
+        override fun asLceType(): Type<Nothing, Nothing, E> = this
 
         data class Typed<out T> @PublishedApi internal constructor(
             override val value: T
@@ -102,26 +102,26 @@ sealed class Type<out L, out C, out E> : LCE<L, C, E> {
         }
     }
 
-    data class Content<out T> @PublishedApi internal constructor(
-        val value: T
+    data class Content<out C> @PublishedApi internal constructor(
+        val value: C
     ) :
-        Type<Nothing, T, Nothing>(),
-        LCE<Nothing, T, Nothing>,
-        UCE<T, Nothing>,
-        UCT<T>,
-        LC<Nothing, T>,
-        UC<T>,
-        CE<T, Nothing>,
-        CT<T> {
+        Type<Nothing, C, Nothing>(),
+        LCE<Nothing, C, Nothing>,
+        UCE<C, Nothing>,
+        UCT<C>,
+        LC<Nothing, C>,
+        UC<C>,
+        CE<C, Nothing>,
+        CT<C> {
 
         override fun isLoading(): Boolean = false
         override fun isContent(): Boolean = true
         override fun isError(): Boolean = false
 
-        override fun contentOrNull(): T = value
+        override fun contentOrNull(): C = value
         override fun errorOrNull(): Nothing? = null
         override fun loadingOrNull(): Nothing? = null
 
-        override fun asLceType(): Type<Nothing, T, Nothing> = this
+        override fun asLceType(): Type<Nothing, C, Nothing> = this
     }
 }