+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [0.2.0] - TBD
 - Added LC type.
 - Added `foldLoading` extension method to `LCE`, `UCE` and `UCT` types. 
+- Added `foldContent` extension method to `LCE`, `UCE` and `UCT` types.
 - Added `foldError` extension method to `LCE`, `UCE` and `UCT` types.
 - Added `fromNullable` factory method to `LCE`, `UCE`, `UCT`, `LC`, `UC` types.
 - Added `fromNullable(content) { [Type].loading() }` factory method to all types.